# services/users-service/src/verification/dto/create-verification.dto.ts
import { IsEmail, IsNotEmpty } from 'class-validator';

export class CreateVerificationDto {
  @IsEmail()
  @IsNotEmpty()
  email: string;
}

export class VerifyEmailDto {
  @IsEmail()
  @IsNotEmpty()
  email: string;

  @IsNotEmpty()
  code: string;
}